requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
autopep8==1.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==38.0.3
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1
django-import-export==3.0.1
django-jazzmin==2.6.0
et-xmlfile==1.1.0
fonttools==4.38.0
future==0.18.2
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.1
MarkupPy==1.14
mysqlclient==2.1.1
mysqldump==0.0.10
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
packaging==23.0
Pillow==9.2.0
protobuf==3.20.3
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stripe==4.2.0
svglib==1.4.1
tablib==3.2.1
tinycss2==1.2.1
toml==0.10.2
tqdm==4.64.1
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
whitenoise==6.3.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
